10 Changes
Move “Select a day to see available times” to the top.

Add option to change language to English.

Show a loader/spinner while submitting.

When clicking "Book now", the page scrolls to the middle; it should scroll to the top instead.

The "Confirmed" button should start disabled.

Properly validate all fields.

Add a more descriptive image of the service offered.

“Popular” & “Best Value” tags are barely visible.

Logo does not load in the About section.